+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `gauge_obj:collect()`;
   - `histogram_obj:collect()`;
   - `summary_obj:collect()`;
+  - `metrics.collect()`;
 
 ### Changed
 - Setup cartridge hotreload inside the role